requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.6
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-filter==1.0.4
djangorestframework==3.7.0
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
psycopg2==2.7.3.1
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.1
